Attempting to improve #219
We now have a new setting "Advanced Search". What this does is for each item  in the plex library we will get the ratingKey and perform another API call to get the MetaData for that ratingKey, inside the metadata it contains the "guid", the guid contains either a IMDBId or TheTVDBID. We will then use that to match the request. This does mean that we now need to capture the TvDbId for each TV Request.