Optimized images, Update old compressed image with a new lossless one. (#1514)
* Losslessly optimized the Old branch images

I had used FileOptimizer with the lossless options on the images to reduce their size while keeping the quality. Metadata was stripped but I don't think it was needed. Just optimized old excluding the emby-logo-dark.jpg, see other commits for more on that.

* Remove lossy emby-logo-dark.jpg and replace with matching PNG lossless file optimized with new paths added

Again I had used FileOptimizer with the lossless options on the images to reduce their size while keeping the quality. Metadata was stripped but I don't think it was needed.

I removed the existing JPG as it had bad aliasing and was banding in its color. Replaced it with a lossless png that I redid to match the same jpg and then updated the path info inside the source files to use the png.

* Optimize existing Current branch images

So I had used FileOptimizer with the lossless options on the images to reduce their size while keeping the quality. Metadata was stripped but I don't think it was needed.

* Replace bad aliased image with new updated lossless and replace file path to update to the new one

Again I had used FileOptimizer with the lossless options on the images to reduce their size while keeping the quality. Metadata was stripped but I don't think it was needed.

I removed the existing JPG as it had bad aliasing and was banding in its color. Replaced it with a lossless png that I redid to match the same jpg and then updated the path info inside the source files to use the png.

* Added missing image

Added missing emby-logo-dark.png